+++ app/models/organization.rb
@@ -551,8 +551,9 @@ class Organization < ApplicationRecord
           '5' => { status: 'true', value: 'updated_at_ds', sort_name: true },
           '6' => { status: 'true', value: 'created_at_ds', sort_name: true },
           '7' => { status: 'true', value: 'file_display_name_ss', sort_name: true },
-          '8' => { status: 'true', value: 'collection_resource_title_ss', sort_name: true },
-          '9' => { status: 'true', value: 'associated_file_content_type_ss', sort_name: true }
+          '8' => { status: 'true', value: 'associated_file_file_name_ss', sort_name: true },
+          '9' => { status: 'true', value: 'collection_resource_title_ss', sort_name: true },
+          '10' => { status: 'true', value: 'associated_file_content_type_ss', sort_name: true }
         }
     }.to_json
 
@@ -563,8 +564,9 @@ class Organization < ApplicationRecord
       '3' => { status: 'true', value: 'language_ss', sort_name: true },
       '4' => { status: 'true', value: 'description_ss', sort_name: true },
       '5' => { status: 'true', value: 'file_display_name_ss', sort_name: true },
-      '6' => { status: 'true', value: 'collection_resource_title_ss', sort_name: true },
-      '7' => { status: 'true', value: 'associated_file_content_type_ss', sort_name: true }
+      '6' => { status: 'true', value: 'associated_file_file_name_ss', sort_name: true },
+      '7' => { status: 'true', value: 'collection_resource_title_ss', sort_name: true },
+      '8' => { status: 'true', value: 'associated_file_content_type_ss', sort_name: true }
     }.to_json
     update(file_index_display_column: display_columns_update, file_index_search_column: search_columns_update) if file_index_display_column.blank?
   end
