+++ app/services/aviary/field_management/organization_field_manager.rb
@@ -12,13 +12,6 @@ module Aviary
   module FieldManagement
     # OrganizationFieldManagement Class for Organization Field Management and saving and fetching information
 
-    # BasedFieldManager
-    class BasedFieldManager
-      def sort_fields(fields, sort_by)
-        fields.sort_by { |_key, value| value[sort_by].present? ? value[sort_by] : 999 }.to_h if fields.present?
-      end
-    end
-
     # OrganizationFieldManager
     class OrganizationFieldManager < BasedFieldManager
       include OrganizationHelper
@@ -161,192 +154,6 @@ module Aviary
 
           result
         end
-      end
-    end
-
-    # CollectionFieldManager
-    class CollectionFieldManager < BasedFieldManager
-      def collection_resource_field_settings(collection, type)
-        unless collection.collection_fields_and_value.present? && collection.collection_fields_and_value.resource_fields.present? && collection.collection_fields_and_value.collection_fields.present?
-          org_field = OrganizationFieldManager.new.organization_field_settings(collection.organization, nil, type, 'sort_order')
-          collection_fields = {}
-          if org_field.present?
-            org_field.each_with_index do |(system_name, single_collection_field), index|
-              if single_collection_field['is_default'].to_s.to_boolean? && system_name.present?
-                collection_fields[system_name] = { status: true, tombstone: false, sort_order: index }
-              end
-            end
-          end
-
-          if collection.collection_fields_and_value.present?
-            collection.collection_fields_and_value.update(type.to_s => collection_fields) if collection.collection_fields_and_value[type.to_sym].blank?
-          else
-            obj = CollectionFieldsAndValue.create(collection: collection, organization: collection.organization)
-            obj[type.to_sym] = collection_fields
-            obj.save
-          end
-          sort_fields(collection.collection_fields_and_value[type.to_sym], 'sort_order')
-          collection.collection_fields_and_value.save
-        end
-        collection.collection_fields_and_value
-      end
-
-      def update_field_settings(updated_values, collection, type = 'resource_fields')
-        fields = collection.collection_fields_and_value[type.to_sym]
-        updated_values.each_with_index do |(_index, single_collection_field), _key|
-          single_collection_field.each do |key, value|
-            value = value.to_i if key.include? 'sort'
-            fields[single_collection_field['system_name']][key] = value
-          end
-        end
-        collection.collection_fields_and_value[type.to_sym] = fields
-        collection.collection_fields_and_value.save
-      end
-
-      def update_field_settings_collection(update_information, collection, type = 'resource_fields', sort_order = nil)
-        if collection.class.name == 'ActiveRecord::Relation'
-          collection.each do |single_collection|
-            update_collection_fields(single_collection, type, update_information, sort_order)
-          end
-        else
-          update_collection_fields(collection, type, update_information, sort_order)
-        end
-      end
-
-      def delete_field(collection, type, field_system_name)
-        fields = collection_resource_field_settings(collection, type)
-        fields[type.to_sym].delete(field_system_name)
-        fields.save
-      end
-
-      private
-
-      def update_collection_fields(single_collection, type, update_information, sort_order = nil)
-        collection_fields = collection_resource_field_settings(single_collection, type)
-        fields = collection_fields[type.to_sym]
-        update_information['sort_order'] = sort_order.present? ? sort_order : fields.keys.count + 1
-
-        fields[update_information['system_name']] = update_information.except('system_name')
-        collection_fields[type.to_sym] = fields
-        collection_fields.save
-      end
-    end
-
-    # FieldManager
-    class FieldManager
-      attr_accessor :field_settings, :key
-
-      def initialize(field_settings, key)
-        map_field_settings(field_settings, key)
-      end
-
-      def map_field_settings(field_settings, key)
-        self.field_settings = field_settings
-        self.key = key
-      end
-
-      def info_of_attribute(attribute)
-        field_settings[attribute.to_s] if attribute.present? && field_settings[attribute.to_s].present?
-      end
-
-      def field_key
-        key
-      end
-
-      def field_type
-        field_settings['field_type']
-      end
-
-      def label
-        field_settings['label']
-      end
-
-      def system_name
-        field_settings['system_name']
-      end
-
-      def editor
-        field_settings['editor']
-      end
-
-      def required?
-        field_settings['is_required'].present? ? field_settings['is_required'].to_s.to_boolean? : false
-      end
-
-      def public?
-        field_settings['is_public'].present? ? field_settings['is_public'].to_s.to_boolean? : false
-      end
-
-      def default?
-        field_settings['is_default'].present? ? field_settings['is_default'].to_s.to_boolean? : false
-      end
-
-      def vocabulary?
-        field_settings['is_vocabulary'].present? ? field_settings['is_vocabulary'].to_s.to_boolean? : false
-      end
-
-      def vocabulary_list
-        field_settings['vocabulary']
-      end
-
-      def dropdown_options
-        field_settings['field_configuration'].present? && field_settings['field_configuration']['dropdown_options'] ? field_settings['field_configuration']['dropdown_options'] : []
-      end
-
-      def help_text
-        field_settings['help_text']
-      end
-
-      def repeatable?
-        field_settings['is_repeatable'].present? ? field_settings['is_repeatable'].to_s.to_boolean? : false
-      end
-
-      def field_configuration
-        field_settings['field_configuration']
-      end
-
-      def field_level
-        field_settings['field_level']
-      end
-
-      def sort_order_on_detail_page
-        field_settings['sort_order'].present? ? field_settings['sort_order'].to_i : 999
-      end
-
-      def should_display_on_detail_page
-        field_settings['description_display'].present? ? field_settings['description_display'].to_s.to_boolean? : false
-      end
-
-      def should_display_on_resource_table
-        field_settings['resource_table_display'].present? ? field_settings['resource_table_display'].to_s.to_boolean? : false
-      end
-
-      def should_search_on_resource_table
-        field_settings['resource_table_search'].present? ? field_settings['resource_table_search'].to_s.to_boolean? : false
-      end
-
-      def solr_display_column_name
-        field_settings['solr_display_column']
-      end
-
-      def tombstone?
-        field_settings['is_tombstone'].present? ? field_settings['is_tombstone'].to_s.to_boolean? : false
-      end
-
-      def solr_search_column_name
-        field_settings['solr_search_column']
-      end
-
-      def sort_order_on_resource_table
-        field_settings['resource_table_sort_order'].present? ? field_settings['resource_table_sort_order'].to_i : 999
-      end
-
-      def should_display_on_search_page
-        field_settings['search_display'].present? ? field_settings['search_display'].to_s.to_boolean? : false
-      end
-
-      def sort_order_on_search_page
-        field_settings['search_sort_order'].present? ? field_settings['search_sort_order'].to_i : 999
       end
     end
   end
