+++ app/views/indexes/form/show.html.erb
@@ -21,12 +21,12 @@
                       <h1 id="index_title_heading" class="cursor-pointer" style="text-transform: none;"><%= @file_index.title %></h1>
 
                       <%= simple_form_for @file_index, html: { class: 'd-none', id: "save_file_index_title_form" },
-                        url: update_index_title_path(@file_index.id) do |f| %>
+                        url: update_file_index_partial_path(@file_index.id) do |f| %>
                         <input class="string required form-control" id="edit_file_index_title" required="required" aria-required="true" type="text"
                           value="<%= @file_index.title %>" name="file_index[title]" />
 
-                        <input class="cancel_file_index_title btn btn-secondary" type="button" value="Cancel" />
-                        <input class="save_file_index_title btn btn-primary" type="submit" value="Save" />
+                        <input class="cancel_file_index_title inline_cancel btn btn-secondary" type="button" value="Cancel" />
+                        <input class="save_file_index_title inline_save btn btn-primary" type="submit" value="Save" />
                       <% end %>
                     </div>
                   <% end %>
